fix: edit swaprouter

+++ contracts/interface/IStablePool.sol
@@ -4,24 +4,30 @@ pragma solidity =0.8.10;
 
 interface IStablePool {
     function N_COINS() external view returns (int128);
-    function coins(int128) external view returns (address);
+
+    function coins(uint256) external view returns (address);
+
     function get_virtual_price() external view returns (uint256);
 
     function add_liquidity(
-    // sBTC pool
+        // sBTC pool
         uint256[3] calldata amounts,
         uint256 min_mint_amount
     ) external;
 
     function add_liquidity(
-    // bUSD pool
+        // bUSD pool
         uint256[4] calldata amounts,
         uint256 min_mint_amount
     ) external;
 
-    function remove_liquidity_imbalance(uint256[4] calldata amounts, uint256 max_burn_amount) external;
+    function remove_liquidity_imbalance(
+        uint256[4] calldata amounts,
+        uint256 max_burn_amount
+    ) external;
 
-    function remove_liquidity(uint256 _amount, uint256[4] calldata amounts) external;
+    function remove_liquidity(uint256 _amount, uint256[4] calldata amounts)
+        external;
 
     function exchange(
         int128 from,
@@ -31,10 +37,15 @@ interface IStablePool {
         address receiver
     ) external returns (uint256);
 
-    function calc_token_amount(uint256[2] calldata amounts, bool isDeposit) external returns (uint256);
+    function calc_token_amount(uint256[2] calldata amounts, bool isDeposit)
+        external
+        returns (uint256);
 
-    function calc_token_amount(uint256[3] calldata amounts, bool isDeposit) external returns (uint256);
+    function calc_token_amount(uint256[3] calldata amounts, bool isDeposit)
+        external
+        returns (uint256);
 
-    function calc_token_amount(uint256[4] calldata amounts, bool isDeposit) external returns (uint256);
+    function calc_token_amount(uint256[4] calldata amounts, bool isDeposit)
+        external
+        returns (uint256);
 }
-
